frontend/requirements.txt
flask
requests
python-dotenv
fastapi
uvicorn
pydantic
sqlalchemy
databases

